Make tests safer, eliminate some axios use
Added check to test requiring a GitHub personal access token to skip
the test if the token is not available.

This turned up a linting error: axios on the import blacklist.
Converted use of axios to HttpClientFactory where possible.  The
client produced by a HttpClientFactory does not seem to support
streaming so I did not replace the use of axios when downloading
GitHub release artifacts since it used streaming and those can be
large.  The ghub uses are in exported function signatures so those
could not be converted without breaking compatibility.  The use in the
test seems fine.