Make tests safer, eliminate some axios use
Added check to test requiring a GitHub personal access token to skip
the test if the token is not available.

This turned up a linting error: axios on the import blacklist.
Converted use of axios to HttpClientFactory where possible.

The client produced by a HttpClientFactory does not seem to support
streaming so I did not replace the use of axios when downloading GitHub
release artifacts since it used streaming and those can be large.  I did
discover and fix a bug in the downloadFileAs function.  The approach
used only works for public repos and only if you do not send an auth
header.  Because of a bug merging the options, the auth header was not
being used, so it worked.  Fixing the merge bug caused the download to
fail.  I removed the use of the auth header, and the incorrect
Content-Type header, and left a comment about the limitations of the
approach.

The ghub uses are in exported function signatures so those could not be
converted without breaking compatibility.  The use in the test seems
fine.

Update TypeScript packages and fix some npm audit issues.